second

+++ account/views.py
@@ -1,4 +1,4 @@
-from django.http import JsonResponse
+from django.http import HttpRequest, HttpResponse, JsonResponse
 from rest_framework import generics
 from types import GenericAlias
 from django.urls import reverse_lazy
@@ -6,7 +6,7 @@ from rest_framework.response import Response
 from rest_framework import status
 from rest_framework.views import APIView
 from account.serializers import UserRegrstrationSerializer, UserLoginSerializer, UserDataViewSerializer, UserChangePasswordSerializer, SendPasswordResetEmailSerializer, UserPasswordResetSerializer, UserUpdateDetailsSerializer, AllUserDataViewSerializer
-from .middlewere import ExampleMiddleware
+from account.middlewere import ExampleMiddleware
 from .serializers import RoleSerializer, PermissionSerializer, RolePermissionSerializer, UserPermissionSerilizer
 from django.contrib.auth import authenticate
 from rest_framework_simplejwt.tokens import RefreshToken
@@ -18,6 +18,8 @@ from django.utils.encoding import force_str
 from django.contrib.auth.tokens import default_token_generator
 from .models import Role
 from .models import Permission, RolePermission, UserPermission
+from django.utils.decorators import method_decorator
+from django.views.decorators.csrf import csrf_exempt
 
 # generate token manually
 
@@ -36,17 +38,7 @@ class RoleListCreateView(generics.ListCreateAPIView):
     
     permission_classes = [permissions.IsAuthenticated]
 
-    is_authorized_middleware = ExampleMiddleware()
-
-    # def get(self, request, *args, **kwargs):
-    #     # Check if the user is authorized
-    #     if not self.is_authorized_middleware.has_permission(request):
-    #         # Return an error response if the user is not authorized
-    #         return JsonResponse({"error": "You do not have permission to access this resource."}, status=403)
-        
-    #     # Continue with your view logic if the user is authorized
-    #     # For example:
-    #     return JsonResponse({"message": "Success!"})
+   
 
     queryset = Role.objects.all()
     serializer_class = RoleSerializer
@@ -212,26 +204,101 @@ class UserLoginView(APIView):
                 return Response({'errors': {'non_field_errors': ['email not validate or email or pwd are not validate']}}, status=status.HTTP_400_BAD_REQUEST)
 
 
+# class UserDataView(APIView):
+#     permission_classes = [permissions.IsAuthenticated]
+
+#     def get(self, request, *args, **kwargs):
+#         django_request = HttpRequest()
+#         django_request.method = request.method
+#         django_request.GET = request.query_params
+#         django_request.POST = request.data
+#         django_request.user = request.user
+#         # Set this attribute to True to disable CSRF checks
+#         django_request._dont_enforce_csrf_checks = True
+
+#         middleware = ExampleMiddleware(get_response=self.dispatch)
+#         response = middleware(django_request)
+#         print(response)
+#         if response is not True:
+#              return response
+    
+#         serializer = UserDataViewSerializer(request.user)
+#         return Response(serializer.data, status=status.HTTP_200_OK)
+      
 class UserDataView(APIView):
+
     permission_classes = [permissions.IsAuthenticated]
 
     def get(self, request, *args, **kwargs):
-        serializer = UserDataViewSerializer(request.user)
-        return Response(serializer.data, status=status.HTTP_200_OK)
 
+        view_name = 'dataview'
+
+        django_request = HttpRequest()
+        django_request.method = request.method
+        django_request.GET = request.query_params
+        django_request.POST = request.data
+        django_request.user = request.user
+        # Set this attribute to True to disable CSRF checks
+        django_request._dont_enforce_csrf_checks = True
+
+        middleware = ExampleMiddleware(get_response=self.dispatch)
+        response = middleware(django_request,view_name=view_name)
+        # middleware(django_request,view_name=view_name)
+
+        # if not isinstance(response, HttpResponse):
+            # Middleware returned True, indicating user has permission
+        if response.status_code == 200:    
+            serializer = UserDataViewSerializer(request.user)
+            return Response(serializer.data, status=status.HTTP_200_OK)
+        else:
+            return response
+            # Middleware returned an error response
+    
+  
 
 class AllUserDataView(APIView):
 
-    def get(self, request):
-        users = User.objects.all()
-        serializer = AllUserDataViewSerializer(users, many=True)
-        return Response(serializer.data, status=status.HTTP_200_OK)
+    permission_classes = [permissions.IsAuthenticated]
+    
+
+    # def get(self, request):
+    #     django_request = HttpRequest()
+    #     django_request.method = request.method
+    #     django_request.GET = request.query_params
+    #     django_request.POST = request.data
+    #     django_request.user = request.user
+    #     # Set this attribute to True to disable CSRF checks
+    #     django_request._dont_enforce_csrf_checks = True
+
+    #     middleware = ExampleMiddleware(get_response=self.dispatch)
+    #     response = middleware(django_request)
+    #     print(response)
+    #     if response is not True:
+    #          return response
+        
+    #     users = User.objects.all()
+    #     serializer = AllUserDataViewSerializer(users, many=True)
+    #     return Response(serializer.data, status=status.HTTP_200_OK)
 
 
 class UserUpdateDetailsView(APIView):
     permission_classes = [permissions.IsAuthenticated]
 
     def put(self, request, format=None):
+        django_request = HttpRequest()
+        django_request.method = request.method
+        django_request.GET = request.query_params
+        django_request.POST = request.data
+        django_request.user = request.user
+        # Set this attribute to True to disable CSRF checks
+        django_request._dont_enforce_csrf_checks = True
+
+        middleware = ExampleMiddleware(get_response=self.dispatch)
+        response = middleware(django_request)
+        print(response)
+        if response is not True:
+             return response
+        
         seriazer = UserUpdateDetailsSerializer(
             data=request.data, context={'user': request.user})
         if seriazer.is_valid(raise_exception=True):
@@ -271,6 +338,20 @@ class UserDeleteView(APIView):
     permission_classes = [permissions.IsAuthenticated]
 
     def delete(self, request, format=None):
+        django_request = HttpRequest()
+        django_request.method = request.method
+        django_request.GET = request.query_params
+        django_request.POST = request.data
+        django_request.user = request.user
+        # Set this attribute to True to disable CSRF checks
+        django_request._dont_enforce_csrf_checks = True
+
+        middleware = ExampleMiddleware(get_response=self.dispatch)
+        response = middleware(django_request)
+        print(response)
+        if response is not True:
+             return response
+        
         user = request.user  # Get authenticated user
         user.delete()
         return Response({'msg': 'User deleted successfully'}, status=status.HTTP_200_OK)
